Changed the coordinates to 'float'
(cherry picked from commit 63604370c115676b751db9e431a727d7d13adeb6)

+++ core/src/de/tum/cit/ase/maze/Enemy.java
@@ -11,7 +11,7 @@ public class Enemy extends GameObject{
      * @param x The initial x-coordinate of the Enemy.
      * @param y The initial y-coordinate of the Enemy.
      */
-    public Enemy(int x, int y) {
+    public Enemy(float x, float y) {
         super(x, y);
     }
